Extract BRK config classes in per-collection files
