+++ src/gobexport/exporter/config/brk/csv_format.py
@@ -1,0 +1,19 @@
+class BrkCsvFormat:
+    def _prefix_dict(self, dct: dict, key_prefix: str, val_prefix: str):
+        return {f"{key_prefix}{key}": f"{val_prefix}{val}" for key, val in dct.items()}
+
+    def _add_condition_to_attrs(self, condition: dict, attrs: dict):
+        return {k: {**condition, 'trueval': v} for k, v in attrs.items()}
+
+    def show_when_field_notempty_condition(self, fieldref: str):
+        return {
+            "condition": "isempty",
+            "reference": fieldref,
+            "negate": True,
+        }
+
+    def show_when_field_empty_condition(self, fieldref: str):
+        return {
+            "condition": "isempty",
+            "reference": fieldref,
+        }
